test(utils): drop core dependency and use regular Babel plugin for testing
