+++ packages/utils/package.json
@@ -13,8 +13,5 @@
   "dependencies": {
     "@ast-decorators/typings": "^0.1.0",
     "@babel/types": "^7"
-  },
-  "devDependencies": {
-    "@ast-decorators/core": "^0.1.0"
   }
 }
