+++ packages/utils/tests/fixtures/createPropertyByPrivacy/none/input.ts
@@ -1,6 +1,4 @@
-import foo from '../decorators';
-
 // @ts-ignore
 class Foo {
-  @foo public bar?: string;
+  public bar?: string;
 }
